Remove full brake workaround and added metadrive to docs
