+++ src/scenic/simulators/metadrive/__init__.py
@@ -2,7 +2,7 @@ metadrive = None
 try:
     import metadrive
 except ImportError:
-    pass
+    raise ImportError("Metadrive is required. Please install the 'metadrive' package.")
 if metadrive:
     from .simulator import MetaDriveSimulator
 del metadrive
