--- tests/simulators/metadrive/test_basic.py
@@ -1,15 +1,0 @@
-import pytest
-
-from tests.utils import pickle_test, sampleScene, tryPickling
-
-
-def test_basic(loadLocalScenario):
-    scenario = loadLocalScenario("basic.scenic", mode2D=True)
-    scenario.generate(maxIterations=1000)
-
-
-@pickle_test
-@pytest.mark.slow
-def test_pickle(loadLocalScenario):
-    scenario = tryPickling(loadLocalScenario("basic.scenic", mode2D=True))
-    tryPickling(sampleScene(scenario, maxIterations=1000))
